moved persisted array function to it's own file in hooks
